requirements.txt
flask
requests
pyjwt
flask-limiter
werkzeug

pyopenssl
cryptography
